1. made AddABook reusable so one component can add a book to the userbib or saved
